switch to using xctool for unit testing

+++ Rakefile
@@ -2,8 +2,6 @@ require 'rubygems'
 require 'pathname'
 require 'cucumber/rake/task'
 
-require 'xcodebuild'
-
 ROOT_DIR = Pathname.new(__FILE__) + ".."
 PROJECT_DIR = ROOT_DIR + "Run2Bart"
 ARTIFACTS_DIR = ROOT_DIR + "artifacts"
@@ -14,29 +12,11 @@ def move_into_artifacts( src )
 end
 
 task "clear-artifacts" do
-  ARTIFACTS_DIR.rmtree
+  ARTIFACTS_DIR.rmtree if ARTIFACTS_DIR.exist?
+  ARTIFACTS_DIR.mkpath
 end
 
 require_relative 'travis/rake_tasks'
-
-namespace 'xcode' do
-
-  XcodeBuild::Tasks::BuildTask.new('sim-debug') do |t|
-    t.invoke_from_within = './Run2Bart'
-    t.configuration = "Debug"
-    t.sdk = "iphonesimulator"
-    t.formatter = XcodeBuild::Formatters::ProgressFormatter.new
-  end
-
-  XcodeBuild::Tasks::BuildTask.new('unit-tests') do |t|
-    t.invoke_from_within = './Run2Bart'
-    t.workspace = "Run2Bart.xcworkspace"
-    t.configuration = "Debug"
-    t.sdk = "iphonesimulator"
-    t.scheme = 'CommandLineTests'
-    t.xcconfig = 'command-line-tests.xcconfig'
-  end
-end
 
 namespace :frank do
   desc "build a Frankified version of the app for testing purposes"
@@ -75,7 +55,9 @@ task 'quit-simulator' do
 end
 
 desc "run all application unit tests"
-task 'unit-tests' => ['quit-simulator','xcode:sim-debug:cleanbuild','xcode:unit-tests:build']
+task 'unit-tests' do
+  sh %Q|xctool -workspace #{PROJECT_DIR+"Run2Bart.xcworkspace"} -scheme Run2Bart test|
+end
 
 task :default => "unit-tests"
